Intermediate commit.
Heavy changes in message/radio implementation to support multi-mode protocol.

The main virtual channel will be dedicated to small rapid exchanges. When a
node receives the beacon of another, and can infer that there is information to
be shared, that node will send a special beacon to request a 1 on 1 conversation
that takes place in a "private virtual channel".

While 2 nodes are sharing a conversation in the PVC, other nodes may also
pair off to do the same. When the exchange is finished, they will return to
the main channel and repeat the process.

This is early experimental code, and is subject to more large changes.

Complete documentation will be written when things settle down.

Signed-off-by: Scott Carlson <scottrcarlson@gmail.com>